Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > nDiscoverIdentityCount only applies in Explicit Contract.
> 
> I don't think that's the intent of the counter, given it's referenced in PE_SRC_VDM_Identity_Request State which is 100% definitely running before an explicit contract is in place.  
> 
> The timer only applies to while we're in an explicit contract, and I clarified with the WG there's no specific discovery timing requirement before a contract (I just choose to still space out requests, because spamming too quickly seems like a poor plan).
> 
> For the plan forward here, I'm going to try searching the PD archives to see if I can find discussion as to what prompted the retry paragraph in the specification revision section and see if it has clarification about circumstances assumed and intent.

From responses I got back, sounds like we specifically only down-rev when we've received a GoodCRC from the partner.  To not receive a GoodCRC would indicate a cable which is especially violating spec since they must GoodCRC all messages regardless of revision.  

For whether there are other reasons to expect a GoodCRC with no response, I'm not coming up with any outside a repeat msgid that would be valid.  If another message came in, we'd go back to ready without the state timing out.  If a cable really is PD 3.0, we should have collision avoidance working in our favor.  If a cable is taking a while to power on, we wouldn't expect a GoodCRC to come in. 

For the soft reset idea, I'm not certain how necessary it will be.  This is either our first communication with the cable, in which case we should be in sync.  Or we Vconn swapped, in which case we already cable soft reset as a part of that process.  I suppose we could have things out of sync if the port partner erroneously communicated with the cable when it shouldn't have?

Patch-set: 1