step back to 0.3.26 state
The MP branch got ahead of itself, we need to publish an intermediate
release before 5.4. Let's bring the branch back to 0.3.26 state so
changes necessary for the intermediate release can be made cleanly.
Once the release is cut, this patch will be reverted to bring the
branch back to the top.

BRANCH=none
BUG=none
TEST='git diff 831c53f' returns no output.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: If47166b97d1f0faf1c315982827e2c2d5ce00c78