Update patch set 4
Patch Set 4:

(2 comments)

Is there a specific section in this that handles the USB portion that was not functioning? Apologies - since I'm new to this code. I think it's the fault handling part that was causing the USB issues? But was there a specific area that was the hotspot?

Patch-set: 4
CC: Gerrit User 1390635 <1390635@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1380689 <1380689@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>