Update patch set 5
Patch Set 5: Verified+1

(5 comments)

> Patch Set 4:
> 
> (2 comments)
> 
> Is there a specific section in this that handles the USB portion that was not functioning? Apologies - since I'm new to this code. I think it's the fault handling part that was causing the USB issues? But was there a specific area that was the hotspot?

This was to keep USB memory sticks in a state they can be located after reboots and power swaps. The problem is the answer required a number of things to be fixed as a unit in order for this to work.

Patch-set: 5
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1