raa489000: Add support for DVC
This commit adds support for DVC, Dynamic Voltage Compensation, which
is supported by the RAA4890000.  In DVC mode, the battery can be
charged through an auxiliary RAA489000 operating as a Voltage
Regulator.  DVC mode is used to autonomously fine tune the charge
current without involvement from the system.  Therefore, the main OCPC
PID loop should not be needed.

BUG=b:155225932
BRANCH=None
TEST=Build and flash waddledoo, verify that battery can be charged
from the sub-board.
TEST=`make -j buildall`

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ie6af4fa4abcda2c8695919e30d231d812d7171b5