gsctool: print sensible date when listing flash log contents
Displaying epoch values in the flash log entries does not help the end
user to synchronize the flash log events with other logs generated on
the device.

This patch changes the log output format to display the device time
zone in the first line and then all present log entries with timestamp
in 'dd:mm:yy hh:mm:ss' format.

BUG=b:153764696
TEST=ran the new gsctool image on a Chrome OS device:

    $ gsctool -a -L
    Log time zone is PST
    Dec 31 69 16:00:01 : 00
    May 06 20 21:20:41 : 09 01
    ...
    May 08 20 12:09:21 : 09 00
    May 08 20 12:09:22 : 09 03
    May 08 20 12:09:23 : 09 06

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I3fca12e1679fbdd9e0e168606014e84c89c42402