VCONN: Request Vconn swap if power state transitions above G3
To communicate with cable plug VCONN is required and the VCONN is
typically provided by 5V rail. When the DUT is below S5, 5V rail is
down & VCONN sourcing is disabled. Hence need to re-enable the
VCONN sourcing when the power state transition happens from G3 to higher state.

BUG=b:156174498
BRANCH=none
TEST=Connect TBT device in G3 behind DP port connected to external monitor.
     Observed TBT alt mode after wakeup

Signed-off-by: madhusudanarao amara <madhusudanarao.amara@intel.corp-partner.google.com>
Change-Id: I011b33f2f7cd56669f84781c5d1ad014e00387ef
Signed-off-by: madhusudanarao amara <madhusudanarao.amara@intel.corp-partner.google.com>