WIP: Battery: Implement safe discharge system
Currently, the EC has only two behaviors when the system is left idle.
One is to hibernate and the other is to cut off the battery. And these
are determined at compile time.

If the system hibernates, EC will not have a chance to cut off the
the battery before the state of charge reaches critical low. If the
system is cut off, it requires an AC adapter to wake up. So, neither
behavior is ideal.

This patch introduces the safe discharge system. Given the number
of hours to survive as a target, it tries to choose the better
behavior.

For example, if the state of charge is high, it will hibernate the
system because the target can be met before the battery completely
drains. If the state of charge is low, it will keep the EC stay up
so that it can cut of the battery.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:152431365
BRANCH=none
TEST=TBD

Change-Id: I1470b13203f3653ae0e495cd5ec8ed05f3c5102f