tcpmv2: hard reset AutoDischarge enable only when needed
The allow unattach path is common code in PE and
it gets called even if we are not doing Hard Reset.
This can change AutoDischarge to be incorrect during
a PR Swap.  So guard this code to only being performed
if we are actually in a Hard Reset.

BUG=none
BRANCH=none
TEST=check Hard Reset and PR Swap

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Ie7e9084d41432d6c449aaaa189b3a55d0dd7f780