Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> Sorry, haven't had a chance to dig into this more on Zork yet.
> What confuses me is the power button. Is that currently broken on ToT on all boards with Vivaldi enabled?

No, the thing is that (I was given to understand) that the power button is supposed to be non-matrixed (i.e. not part of the keyboard). This seems true for most of the systems, but apparently not for Zork maybe. To fix your power button issue, you will need:

1) CL in https://b.corp.google.com/issues/155941390#comment19
2) This CL
3) One more CL that I suspect I may have to come up for coreboot, but for that, I first need the info asked in https://b.corp.google.com/issues/156201575#comment4 (most importantly, the evtest results when power button is pressed).

Patch-set: 1