Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> This ought to work.

Yup, seems to.  It's about the only way I could think of clearing things while always keeping the TCPC interrupt servicing at highest priority.  With a DRP dongle in C1 and a print in the check_c[0,1]_line for when the line is low, we seem to get this condition about 3 times per plug-in.  We get it sometimes on C0 as well on charger un-plug.

Patch-set: 1
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>