syv682x: fix chip init sequence
This fixes a failure mode in the syv682 chip init sequence. Resetting
the chip registers using the RST_REG sets the over-voltage threshold
to 6V and at the same time enables the high voltage path. It is not
unusual for the high voltage channel to carry a high voltage like 15V,
so we get an OVP interrupt before finishing our chip init sequence.

BRANCH=none
BUG=b:156585531
TEST=no more VBUS OVP interrupts

Change-Id: Iab19012d390e0c5dd8f2cb726ac45cd14732c6f8
Signed-off-by: Caveh Jalali <caveh@chromium.org>