TCPMv2: Rearrange DP setup to use a host command
Provide a host command to direct the PE to send SVDMs via a DPM request.
(Right now, the DPM code sets up the SVDM before invoking the host
command to trigger it. This may change.) Handle that DPM request in the
PE. Also in the PE, invoke the host command (via the DPM) after
discovery completes.

Essentially detach PE_DO_PORT_DISCOVERY from the rest of the PE state
machine in preparation for deleting it.

Create separate modules representing DisplayPort (which decides whether
and how we can set up the DP alt mode) and the Device Policy Manager
(which decides whether we should). The end goal is to have essentially
all code with DP-specific decisions or information in the DP module and
all code with DPM-specific decisions or information in the DPM module.
This will allow the PE to map as directly as practical to its behaviors
defined in the PD spec.

This change paves the way for allowing the AP to drive mode entry.

BUG=b:155890173
TEST=Attach DP dongle; observe discovery and DP setup via Twinkie
BRANCH=none

Change-Id: Ie63a2e62d1ac6178722477dc53244898a04ef92f
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>