TCPMv2: Rearrange DP setup to use a host command
Provide a host command to direct the PE to send SVDMs (eventually; right
now it just triggers all of DP setup at once) via a DPM request. Handle
that DPM request in the PE. Also in the PE, invoke the host command
after discovery completes.

This change paves the way for allowing the AP to drive mode entry.

BUG=b:155890173
TEST=Attach DP dongle; observe discovery and DP setup via Twinkie
BRANCH=none

Change-Id: Ie63a2e62d1ac6178722477dc53244898a04ef92f
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>