Update patch set 6
Patch Set 6:

(2 comments)

PTAL

Actually, there is still something missing: as presented, the AP RO hash space could be programmed at any time unless it is already programmed. We don't want the devices already out in the field to allow programming this space.

Need to think how to enforce the policy that this could be programmed only in the factory or during RMA, let's do it in a separate patch though.

Patch-set: 6