TCPMv2: Check whether a message interrupted swap send
An outgoing message may be discarded when an incoming message comes in.
When this happens, detect the condition and return to ready to process
that message.  This commit adds the check into the data role and vconn
swap functions, both of which will spin forever if they don't see
TX_COMPLETE.

BRANCH=None
BUG=b:156758374,b:155229288
TEST=with Ripto, Moshi is able to connect, process messages normally,
and return to a ready state when connection is steady

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I062509c7a89f20ec59e138d951aca73794390bbe