Update patch set 2
Patch Set 2:

AutoDischargeDisconnect is not supposed to be disabled the way I was doing it. The whole idea of this is to allow the Disconnect to be detected and rapidly discharge.  By shutting off the AutoDischargeDisconnect where it was caused the rapid discharge to slow down.  It is better to use the Safe0V alert to detect this needs to happen and disable at that point.  This also requires entering all attached states to blindly enable.

Patch-set: 2
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0