Update patch set 2
Patch Set 2: Code-Review+2

Nit: Nominally, the comments for the affected functions should now be in mpu_private.h, since they have external linkage. This is somewhat of a special case, but putting the comments in the header file might be a little less surprising for future debuggers of the tests that use these functions. In any case, mpu_disable and friends should have comments now that they will be referenced in tests, even if they are very short.

I'm slightly skeptical of exposing interfaces for the sole purpose of writing tests against them, but I'm not familiar enough with the structure of this part of the code to have a strong opinion.

LGTM modulo nits about comments.

Patch-set: 2
Reviewer: Gerrit User 1305927 <1305927@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2