EFS2: Jump to RW when system shuts down
Currently, when the system shuts down and EC is still in RO, it'll
reset with AP_OFF reset flag. This was for allowing the system to
exit the recovery mode when the power button is pressed when the
recovery screen is displayed.

This behavior causes AP reset in the recovery mode to be stopped.

This patch makes RO jump to RW instead of resetting. A new reset
flag, AP_IDLE, is introduced. It makes RW not automatically press
the power button.

Currently, when RW starts, it forces the chipset to start from G3
unless AP is already powered on. This behavior is based on the
assumption that sysjump happens only when AP is up.

With EFS2, it's no longer the case because EC jumps while AP is
off. AP may be resetting (i.e. s0->s5->s0) or may be shutting
down (s0->s5).

This patch makes RW initialize the chipset state to S5 if the AP
is already powered on. It allows the chipset to sequence up or
down but won't force it to go either way.

BUG=b:154778457
BRANCH=none
TEST=Verify firmware_CorruptBothFwSigAB passes.
TEST=Verify test_that suite:faft_bios passes.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I38588fb6b355f7749dd95ab7d64ffe5a9473077c