EFS2: Jump to RW when system shuts down
1. Use AP_IDLE instead of AP_OFF to restore AP power state

2. Call verify_and_sysjump from shutdown hook instead of system_reset

Currently, when the system shuts down and EC is still in RO, it'll
reset with AP_OFF reset flag. This was for allowing the system to
exit the recovery mode when the power button is pressed when the
recovery screen is displayed.

This behavior causes AP reset in the recovery mode to be stopped.

This patch makes RO jump to RW instead of resetting. A new reset
flag, AP_IDLE, is introduced. It makes RW not automatically press
the power button.

3. Add EC_RESET_FLAG_AP_IDLE

4. Don't check reset_flags to detect sysjump

5. Initialize chipset state to S5 based on power signal after sysjump

Currently, when RW starts, it forces the chipset to start from G3
unless AP is already powered on. This behavior is based on the
assumption that sysjump happens only when AP is up.

With EFS2, it's no longer the case because EC jumps while AP is
off. AP may be resetting (i.e. s0->s5->s0) or may be shutting
down (s0->s5).

This patch makes RW set the chipset state to S5 if the corresponding
power signals are on.

BUG=b:154778457
BRANCH=none
TEST=Verify test_that suite:faft_bios passes.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: If08f3bb7f89c2407e4c5207ade820085f65ea9ac