EFS2: Jump to RW when system shuts down
Currently, when the system shuts down and EC is still in RO, it'll
reset with AP_OFF reset flag. This was for allowing the system to
exit the recovery mode when the power button is pressed when the
recovery screen is displayed.

This behavior causes AP reset in the recovery mode to be stopped.

This patch makes RO jump to RW instead of resetting. A new reset
flag, AP_IDLE, is introduced. It makes RW not automatically press
the power button.

BUG=b:154778457
BRANCH=none
TEST=Verify firmware_CorruptBothFwSigAB passes.
TEST=Verify test_that suite:faft_bios passes.

Change-Id: I0aa3c60b36f5bc729af33898c932f5ffcfbf08c3