Update patch set 6
Patch Set 6:

(3 comments)

> Patch Set 6:
> 
> (3 comments)
> 
> Implementation looks ok. What's the plan for testing?

Thanks for the review! As mentioned in my reply, "The analysis that I've done is simply taken a look at the voltage waveform on an oscilloscope to see that it converges with minimal oscillation. I've also done this in the state where the SoC is off as well as when the SoC is on and therefore the load is more dynamic." I've also tried charging while going between those two states.

I've done this from a low battery state all the way to full to ensure I cover both CC and CV stages of charging.

Patch-set: 6
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>