Update patch set 1
Patch Set 1: Code-Review+1

This matches the code in the bug, and if it works in the correct orientation, I guess that should prevail. However, the comment in the bug suggests that only the Y and Z orientations need to change, and that seems consistent with the pictures in the bug. Should the X axis be flipping?

Patch-set: 1
Reviewer: Gerrit User 1305927 <1305927@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1