EC: Expand recovery button handling to allow second button
Puff has a second dedicated recovery button input, so
expand the common recovery button handler to include
a second recovery button signal. The signals are treated
as an OR, so that either signal will indicate recovery.

The reason a second input is required is that the signal from
the H1 for recovery is masked off if the power button is pressed,
so that recovery mode of pressing both power and recovery is not
possible.

BUG=b:157181336
TEST=Patched into puff, verified recovery mode.
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Cq-Depend: chromium:2208342
Change-Id: I476b2e8b1f5ec0f91f39e9a568dac4f15a45438f