x86-power: Don't force G3 when RW starts
Currently, when RW starts, it forces the chipset to start from G3
unless AP is already powered on. This behavior is based on the
assumption that sysjump happens only when AP is up.

With EFS2, it's no longer the case because EC jumps while AP is
off. AP may be resetting (i.e. s0->s5->s0) or may be shutting
down (s0->s5).

This patch makes RW initialize the chipset state to S5 if the AP
isn't already powered on. It allows the chipset to sequence up or
down but won't force it to go either way.

BUG=none
BRANCH=none
TEST=Run test_that suite:faft_bios on Puff.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I53a036b05f1f6083cb51be4cb0217e13233ee1b9