Update patch set 2
Patch Set 2: Verified+1

So I think this is the more correct path, so that if we select not to restore the
state from BBRAM, it will clear the explicit contract flag and also restart the TCPC.
Unfortunately I don't have a Kaisa, so it is impossible to test.
I am not familiar enough with the PD code to confirm how to know that the
restart sequence will be fully correct in the RO/RW, and I am presuming that
the restart code should work as intended.

Patch-set: 2
Reviewer: Gerrit User 1281040 <1281040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1