Update patch set 4
Patch Set 4: Code-Review+2

> > I'm slightly skeptical of exposing interfaces for the sole purpose of writing tests against them, but I'm not familiar enough with the structure of this part of the code to have a strong opinion.
> > 
> > LGTM modulo nits about comments.
> 
> Re: the comments, I tend to prefer keeping comments "close to the code" to prevent them from getting out of sync with the implementation. Most editors/IDEs (e.g., go/clion-for-chromeos#automatically-add-doxygen-arguments) and the generated doxygen docs themselves correctly show doxygen comments for functions no matter where the documentation is.

I was thinking of the Google style guide, which requires that function declarations should be commented for the benefit of their consumers. However, this is not mentioned in Linux style and is merely common in the EC codebase. Since there are not intended to be any non-test consumers of these functions, do as you will.

Patch-set: 4
Reviewer: Gerrit User 1305927 <1305927@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2