Update patch set 2
Patch Set 2: Code-Review-1

I think this would work for puff, but I'm a little hesitant about introducing this config as I think the case I mentioned earlier still stands.

In my opinion, if the hardware is designed in such a way that an EC reset also resets the Type-C port, then during puff's board init, it should clear the BBRAM flag (checking the reset cause of course).

My worry is that if someone were to use this CONFIG_* option and the hardware is _not_ designed in that way, the explicit contract still stands, yet the FW will think that it does not exist. This would cause the board to issue a HardReset to the port partner waiting for a SourceCap that may never come.

Patch-set: 2
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1