system: Wait chipset shutdown complete; then call the pending reboot
The pending reboot command is currently triggered in the
CHIPSET_SHUTDOWN hook.

The CHIPSET_SHUTDOWN hook is triggered before the power rails
are removed. The hook is executed in the CHIPSET task context,
meaning the chipset is still ON.

Should wait for the power sequence transition to S5/G3, in a
different task (HOOKS) context; then perform the pending reboot.

Note that on x86 boards, the power sequence is not fully
controlled by EC (EC just listens to the power rails and maps to
a power state). The existing way may work as the chipset is
shutting down in parallel.

BRANCH=None
BUG=b:156981868
TEST=Used an AP image which issues an EC_REBOOT_COLD with the
ON_AP_SHUTDOWN flag. Checked EC console that the chipset power state
completely to S5 and then the reboot command was performed.
TEST=Ran "ectool reboot_ec cold at-shutdown" and "shutdown -H now".

[17.563033 power state 8 = S3->S5, in 0x0005]
[17.581748 C0 Exiting DP mode]
[17.582969 PD:S3->S5]
[17.586942 Gyro ODR: 0 - roundup 0 from config 0 [AP0]]
[17.588071 set_pmic_pwron(0)]
[17.658629 set_system_power(0)]
[17.669709 power shutdown complete]
[17.670241 power button released in time]
RTC: 0x5eb94f13 (1589202707.00 s)
[17.670942 power state 1 = S5, in 0x0005]
Reboot at shutdown: 4

Change-Id: Ic051f2c613629459fe7868853d9a4ad694d5aaf8
Signed-off-by: Wai-Hong Tam <waihong@google.com>