system: Don't check reset_flags to detect sysjump
Currently, system_common_pre_init checks if reset_flags is zero or not
to determine whether it's running after sysjump or not.

This used to be a valid condition because reset flags in NVMEM are
cleared by RO after they're read. This is not the case for
CONFIG_POWER_BUTTON_INIT_IDLE because it clears every flag except
AP_IDLE in case the system loses the power again.

This is causing boards with CONFIG_POWER_BUTTON_INIT_IDLE fail to
detect sysjump (because system_common_pre_init sees a non-zero value
in the reset flag).

This patch removes the check. system_common_pre_init will continue to
check the magic value in the jump data. The jump data magic is
always set before sysjump and cleared after sysjump. So, this check
is sufficient and necessary to detect sysjump.

BUG=b:154778457
BRANCH=none
TEST=Verify test_that suite:faft_bios passes.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ib5b8dca02dbf414f5168050e5755c993285bf791