x86 power: Initialize chipset state to S5 after sysjump
Currently, RW initializes the chipset state to G3 and forces the
chipset to shut down unless the AP is already powered on. This
behavior is based on the assumption that sysjump happens only upon
a request from the AP.

With EFS2, it's no longer the case because EC jumps while the AP
is off. AP may be off, resetting (i.e. s0->s5->s0), or shutting
down (s0->s5).

This patch makes RW set the chipset state to S5 if the corresponding
power signals are on.

BUG=b:154778457
BRANCH=none
TEST=Verify test_that suite:faft_bios passes.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I65cbb6c6e7f8a01e80d83a74e376ceb9628b9789