EFS2: Verify and jump when system shuts down in RO
Currently, when the system shuts down in RO, EC resets with AP_OFF
flag. This was for allowing the EC to jump to RW and get ready for
the next normal boot after a system shuts down from the recovery
mode (chromium:1045209).

This behavior causes the reboot from the recovery mode to be stopped
(b:154778457).

This patch makes RO jump to RW instead of resetting with AP_IDLE
flag. After sysjump, the AP can continue to sequence up or down.

BUG=b:154778457, chromium:1045209, b:156694627, b:157077589
BRANCH=none
TEST=test_that suite:faft_bios passes.
TEST=Press power button in recovery screen. System stays in S5.
TEST=Hibernate EC. Press power button once. System boots.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: If08f3bb7f89c2407e4c5207ade820085f65ea9ac