servo: Adds USB SPI helper functions and context states
Adds context states to USB SPI for handling receiving and
transmitting data. These context states allow us to keep track
of data buffers that can not fit within a single USB packet.

A context state has been added to describe the USB packets accepted
by the USB SPI protocol. This structure include union fields for
easily addressing the different packet's fields and to describe
the basic attributes for a USB packet including it's size and
how many bytes are allocated to the header.

Helper functions process the USB packets and aid in the transfer
of data between them and receive and transmission buffers.

BUG=b:139058552
BRANCH=servo
TEST=Tested with Servo Micro reading and verifying flash
    writes are successful in a loop.

Change-Id: I3e0b31f33f442719123b9e897495cad25e9e6ed4