Update patch set 5
Patch Set 5:

I commented on the bug; we should instead increase the i2c bus speed for the TCPC controller. 100kHz is not fast enough to handle and respond to PD message and compliance with spec timing. Increasing the queue size is only a bandaid.

Can we try 400kHzs and revert this CL. I also think adding a comment next to queue depth size that one shoudl trying increase the TCPC bus speed first would be helpful. Thoughts?

Patch-set: 5
CC: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>