ap_ro: add handling of the corrupted hash
This patch closes the AP RO verification loop on the Cr50 side. If the
check is triggered, the valid AP hash is found and the RO contents is
found to not match the hash, the Cr50 will start a periodic hook
keeping the EC in reset which would prevent the Chromebook from
booting.

The only way to determine the state of the device would be connecting
to the Cr50 console and checking the flash log contents. Seeing a
recent log sequence similar to

1590602806:09 01
1590602819:09 00
1590602820:09 00
1590602821:09 03
1590602823:09 06

would indicate that the device is in the failed AP RO check state. The
only way out of this state on a production device is powercycle.

BUG=b:153764696

TEST=verified that erasing or programming AP RO hash when board ID is
     set is impossible.

     Verified proper shutdown in case AP RO has is present and the AP
     RO space is corrupted.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I1029114126a9a79f80385af7bc8d5467738e04ca