waddledee/waddledoo: Mark BC1.2 as always powered
This commit enables the PI3USB9201_ALWAYS_POWERED flag for the
PI3USB9201 BC1.2 chip.  In these designs, these parts are always
powered so there's no need to explicitly enable a 5V rail.  This was
causing Type-A ports to turn on whenever a charger was plugged in.

BUG=b:151688130, b:150814620
BRANCH=None
TEST=Build and flash waddledoo, verify that Type-A ports are not
powered when a charger is plugged in in S5/G3.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ic478c59e96854633a34f510e9494b908ab01ed7c