TCPMv2: Pending PRL Rev3 messages need to stage for transmit
During an initial attach of a ThinkPad Dock Gen2 to
a trembyle, we get into a condition that we have a
pending message.  This message does not go through
the prl_copy_msg_to_buffer function and this leaves
the number of data_objs not getting set to the
messages expected value of 0.  This causes the
ACCEPT message to look like a BIST to the Dock.

Also there is no reason to perform the buffer
initialization and copy if the length is 0, a
control message, so added an early exit for this
case.

BUG=b:157650576
BRANCH=none
TEST=ThinkPad Dock attach should not show BIST message

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I766e6a5e5fb786f5b6c503a76d1721c303044047