Update patch set 1
Patch Set 1:

ubsan does not, asan and msan do.
I'd rather change the condition to __has_feature(address_sanitizer) || __has_feature(memory_sanitizer).

But then, I'm not sure what is the source of the problem. If this code is compiled with MSan, that should be as good (but a lot slower) as intercepting the function.

Patch-set: 1
Reviewer: Gerrit User 1000455 <1000455@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>