TCPMv2: Fix PRS request handling in the DRP SM
On a power role swap request, another flag needs to be set in the SM to
signal that a power role swap is in progress. There is a function that
does this already, but it was not being called as part of the PR Swap
request handling.

The symptom of this was that the PE was being reset on an FRS event
since PD was disabled as part of the state entry in the SM. This
reinitializes the PE and results in a hard reset.

BUG=b:148144711
BRANCH=NONE
TEST=make buildall
TEST=enable TCPMv2 and ensure PRS happens correctly
TEST=enable FRS and TCPMv2, and ensure the PE is not reset when an FRS
occurs.

Change-Id: Id97d114e6e6f4c4f630eb03e51265594c5b7593c
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>