Update patch set 2
Patch Set 2:

I am not sure this is the correct location for this flag to get set.  It should happen as early as possible by the PE when a PR Swap is going to happen for sure.  If you have a path that is not covered in the current ToT from the PE, then I suggest adding a call from there.

This is one of the things that changed on ToT in the last couple of days.

Patch-set: 2