TCPMv2: Aborted PR Swap should not re-enable AutoDischarge
On a recovery path exit a TCPCI I2C FAULT occurs when we
set the CC lines to open/open if AutoDischargeDisconnect
is currently enabled.

The TCPCI I2C FAULT is not a status of the I2C interface
but rather a status of the last TCPCI I2C write performed
with regards to the current state/settings of the TCPCI.
This is usually caused when AutoDischargeDisconnect is
enabled/disabled when it should not be. This change will
make sure this path will not re-enable and that entering
the recovery path will make sure AutoDischarge is disabled.

We will wait to re-enable until the PE engine indicates
the recovery has ended.

BUG=b:157755939
BRANCH=none
TEST=ThinkPad Dock G2 attach should not cause FAULT-0x01

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: If70ccd90b045949e54c131512d1fb294c61dd1dd