TCPMv2: Aborted PR Swap should not re-enable AutoDischarge
On trembyle, the ThinkPad Dock G2 causes an I2C TCPCI
FAULT when we set the CC lines to open/open on error
recovery exit from a PR Swap gone bad.  This makes
This is caused because we have AutoDischargeDisconnect
enabled on failing exit.  This will make sure the
failing exit will not re-enable AutoDischarge

BUG=b:157755939
BRANCH=none
TEST=ThinkPad Dock G2 attach should not cause FAULT-0x01

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: If70ccd90b045949e54c131512d1fb294c61dd1dd