TCPMv2: Aborted PR Swap should not re-enable AutoDischarge
On a recovery path exit (hard reset) a TCPCI I2C FAULT
occurs when we set the CC lines to open/open. This is
caused due to having AutoDischargeDisconnect enabled
on the failing exit.  This change will make sure the
failing exit will not re-enable AutoDischarge.

This AutoDischargeDisconnect exception case does not
appear to have a single point to re-enable
AutoDischarge.  So I had to catch all of the attach
points to make sure AutoDischargeDisconnect is
enabled.  If it is already enabled, it will not
cause harm to the TCPC's operation

BUG=b:157755939
BRANCH=none
TEST=ThinkPad Dock G2 attach should not cause FAULT-0x01

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: If70ccd90b045949e54c131512d1fb294c61dd1dd