Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Benson/Prashant, should we be exiting DP mode when the SoC goes into S0ix/S3? I don't think we have done this before.
> > 
> > Hopefully this would be something the AP did on its own without needing the EC to initiate since the S0ix transition is always graceful (unlike some S5 transitions)
> 
> I don't think it's correct to exit DP mode when the SoC goes into S0iX or S3.
> 
> We should maintain DP alt mode into S0iX or S3 because we have an expectation that the monitor will remain connected and will use VESA DPMS to put the display into a low power mode.
> 
> Exiting DP mode will necessarily mean that on resume, we will register a disconnect and potentially a reconnect on resume, which will cause the Chrome UI to subtract and readd a monitor, potentially causing the user's windows to shift around significantly.
> 
> Also, maintaining the mode will mean that resume time to external display turning on will be faster than the reconnect scenario.

This is what I would have expected, but the issue is not with the suspend, but with the wake-up.
There is no distinction made at the moment between detecting when a connection is still active
in DP mode or not (at the time of wake-up), so the code just kicks off a discovery request, and this
tries to put the connection into DP Alt mode again.
See the attached bug.

It seemed to me that the real fix should be as part of the wake-up, and that it should
check whether Alt mode is on or not, and do not re-negotiate if it is.
But shutting it down on suspend is certainly simpler, and I was keen to hear from folks what the
'right' thing to do is. I'm OK re-working this to be more selective on the wake-up.

However, there is some pressure to get a fix in for dogfood.
Is it acceptable that this goes in now with a TODO, and I will raise another bug
and fix this next week?

Patch-set: 1
Reviewer: Gerrit User 1281040 <1281040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0