TCPCI: Charge Manager analog rp has to set the register
The tcpci_tcpm_select_rp_value caches the rp value but
the charge manager needs to actually set the value in
the register.  This only happens when the PD is not
in an explicit contract.

This was seen on ThinkPad Gen2 Dock connecting to
trembyle.  On first attach the DUT was not correctly
reflecting Analog Rp and the dock detected this after
the PR_Swap was sent and performed Soft and Hard resets

BUG=b:157755939
BRANCH=none
TEST=Verify Rp tracks SinkTxNG

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I5e2cd1bd2c7fffa06f46e1d7a77868f49454c591